Version the minimum protocol version
This change moves the minimum protocol version into our versioned
sets of cipher suite configurations. This means that users will be
able keep a protocol version enabled, even if we disable it by default,
if they choose to use an explicit version for configuration.

Works towards https://github.com/awslabs/s2n/issues/1

As a nice side-effect, this change also removes the dynamoc memory
allocation for the cipher suite preferences and wire formats. If we
disallow changing these, then it's safe to use the versions that are
on the stack.