Implement s2n_config_set_cipher_preferences()
This change implements s2n_config_set_cipher_preferences. Callers can
invoke it in two ways:

    s2n_config_set_cipher_preferences(config, "default");
    s2n_config_set_cipher_preferences(config, "20150202");

the first way will mean that the invoker gets s2n's latest and greatest
cipher suite and protocol version settings.

The second way means that the the invoker keeps a fixed version of
the settings, so that they can evaluate newer settings and update
when ready.