Initial implementation of a timer
This change implements a timer for s2n, which is ultimately intended to be
used for dynamic record sizing. The API here is pretty basic: start a timer
and reset a timer, which also has the effect of returning how many nanoseconds
have elapsed since the timer was last reset (or started).