Remove a dependency on DH_dup()
This change removes the need to call DH_dup() and instead copies the
relevant Big Numbers directly. This makes it possible to use s2n with
BoringSSL.

Although this means relying on the internal structure of the DH struct,
we already rely on this for serialising the parameters. Additionally,
we rely only on the well-known published internals.

This change also tightens up the memory deallocation in some failure
cases. Although this memory is being deallocated anyway in the handshake
functions, it's best to deallocate inline to make it more clear.