Guard against a s2n_negotiate() bad caller
s2n wipes stuffers and adds a non-deterministic delay in the failure
cases, but this logic is contained inside s2n_recv/s2n_read_full_record.

If a caller uses s2n_negotiate() explicitly, but does not call
s2n_close() subsequent to a failed negotiation, we might skip this
logic. This is a bad call pattern, but we should guard against it.