Avoid calling memset with 0 length arguments
Per issue #161 , there are compiler optimizations that mean calling
memset() with a NULL pointer, even when the length argument is 0,
is a bad idea which may lead to undefined behavior.

This change guards two calls to memset() in the s2n code base
that may result in such a call.

fixes #161