Hashes require 9 bytes, not 8, to finalize
Manuel Barbosa from HASLab - INESC TEC, DCC FC Universidade do Porto,
http://haslab.pt/mbb, contacted us to point out the finalizing a SHA or MD5
hash requires 9 bytes of space, rather than 8. The 9 bytes are the length of
the data and an additional 0x80 byte.

This issue means that our CBC verification isn't always running in constant
time.  Ocasionally an additional hash round will be required (if the size of
the record is congruent to the block size mod 8).

Thankfully our randomized delay on error code, which pauses for between 1ms and
10 seconds is more than adequence to ensure that any timing side channel here
is not detectable or exploitable. Nevertheless, it is worth fixing the error.