Specify `s2n_hash_update` and `s2n_hash_digest`.
And factor `setup_initial_hash_state` and `check_final_hash_state` out
of the `s2n_hash_init`, for reuse in the other specs. I wonder if this
abstraction would be possible if the state argument were not always
called `state`? I.e., could the arg name be a parameter?