Use Cryptol and SAW specs from `./specs`.
I replaced the `./cryptol_specs` files with symlinks to the originals
in `./specs`, but I think the README could take care of this by
summarizing which files are of interest.

I also changed the makefile to generate the logs and the `IS_PATCHED`
files in `./tmp`, so that there's less cruft at the top level.

You can run the proofs locally by doing `make` of `make -j` to run all
proofs, or

    saw <proof script>.saw

to run an individual proof.