Disable 1/n-1 CBC message split in server mode
This changes s2n_send to not split CBC messages into 1 byte
and n-1 byte records in server mode. The 1/n-1 split was
identified as a mitigation for the BEAST attack on SSLv3/TLS
1.0 CBC cipher suites.
The motivation for disabling is interoperability. Google Chrome
found that many sites were broken when 1/n-1 splitting was rolled
out [1]. Clients that place strong expectations on the amount/value
of data in the first TLS record received can be broken.

[1] https://www.imperialviolet.org/2012/01/15/beastfollowup.html