Support Openssl 1.1.0 DH usage
This change adds light wrappers around all of our DH access.
After 1.1.0, all fields of the DH struct are opaque. To make things
more complicated, the 1.1.0 approved accessors are NOT available in
earlier versions of Openssl [1]. Openssl version macro logic is
required to know which API to use. This change attempts to minimize
the amount of code that contains such logic.

[1] https://www.openssl.org/docs/man1.1.0/crypto/DH_get0_pqg.html