const: add const to APIs where possible
It's a bit gnarly to make s2n_send() take a const buffer since we
use -Wcast-qual and the buffer functions take a non-const pointer.
This patch uses the trick of an uintptr_t cast to remove constness.

Still, better to do this here in the internals of the function than
to force callers to do similiar things.

Signed-off-by: Anthony Liguori <aliguori@amazon.com>