s2n_map: a basic map implementation
This change adds s2n_map, a very basic, append-only map implementation for s2n.
Updates are append only and are not thread-safe, but once built a map is
considered immutable and thread-safe lookups are possible.

Internally s2n_map is a simple hash map which uses linear-proving go resolve
any collisions. The load-factor is maintained at 50% or below so the worst-case
collision probability on a fully-loaded map is:

    1 probe:  50%
    2 probes: 25%
    3 probes: 12%
    4 probes: 6%
    5 probes: 3%
    6 probes: 1%

and so on. s2n_map "owns" all of its own memory, so when keys and values are
added their contents are duplicated and the caller may release the originals.
The copies in the map are freed when the map itself is freed.

This map is intended for looking up client certificates and other indexing
problems within s2n.

NOTE: for now, s2n_map uses SHA256 as its hash. A later change may implement
CRC32 or murmer as an s2n_hash implementation, for additional speed.