Add support for signature negotiation
This change adds support for negotiated signature types with TLS1.2.
We prefer RSA only for signatures, and the following digest algorithms
in the following order:

    SHA256
    SHA386
    SHA512
    SHA224
    SHA1

MD5 is not supported due to the SLOTH vulnerability and SHA1 and SHA2 are
intentionally de-prioritized for the same reason.

The signature is negotiated during the client hello message, and exchanged
during the server key exchange.