Add "is_available" method to s2n_cipher
The intention of this change is to allow us to query the availability
of a cipher before using it for TLS cipher_suite implementation. For
example there are certain optimized ciphers in the EVP_* interface
that are only available if the host CPU supports Intel AES-NI.

We'll also use this for compatibility with different versions of
libcrypto. As we add more ciphers(i.e. ChaCha20-Poly1305), we'll
need to still operate correctly when compiled against libcrypto 1.0.x

Currently, the only logic we're using to determine availability is
the EVP_* return value. We can harden the logic in the future.

To be added later:
- Cipher suite implementation selection based on the "is_available"
  return value of ciphers.