Avoid data-dependent shift distances
On CPUs without barrel shifters, integer shifts take time dependent on
the distance being shifted.  As a result, it is important to avoid
using such operations in cryptographic code, where they might leak
sensitive information via a timing side channel.

The most recent CPU with which I'm personally familiar on which this
is a problem is the Intel 80286; but while the transistor footprint of
even a 64-bit barrel shifter is insignificant on modern general
purpose CPUs, it is conceivable that some embedded systems will have
sufficiently limited transistor budgets that they shave a few square
microns by using bit-per-cycle shifters.

Since extracting the high bits of a 16+ bit value x - 1 is in any event
a simpler and faster way of mapping bytes 0 -> 255, 1--255 -> 0, use
that rather than the hypothetically dangerous variable shift.