Use EVP Digest Sign functions for p hash in FIPS mode
Migrate the s2n_p_hash implementations to the EVP Digest Sign
APIs in OpenSSL when in FIPS mode. These APIs are required to be
used explicitly to enable a FIPS-capable libcrypto to use MD5
hashes for the PRF.

Use handlers to control the p_hash implementation to use for
TLS connections. Currently, EVP handlers are used when in FIPS mode
and HMAC handlers are used otherwise. Initialize the handlers per-connection.