Add checks for OPENSSL_FIPS flag to control behavior
When the OPENSSL_FIPS flag is defined, the linked libcrypto is a
FIPS-capable library. In that case, there are restrictions on
permitted cipher suites, hash and hmac algorithms, and SSL
protocol versions.

The use of these flags allows s2n to prevent the execution of
crypto that is not available in the FIPS-capable libraries.