Minimize required handshake hashes
Previously s2n kept a running hash of handshake messages for *every*
supported hash algorithm. The handshake hash is used in the PRF and
the optional CertificateVerify message.

This change disables all unneeded hash algorithms after the
{Client, Server}Hello is received [1]. On my test system, this change
reduced handshake processing time by ~5%. In practice, most handshake
time is spent waiting on network round trips and not on-CPU processing.
This change is unlikely to have a significant performance impact
outside of low latency networks.

closes #234

[1] When client authentication is required we delay minimizing handshake
    hashes until the CertificateVerify message is processed.