Add libcrypto argument to integration tests
The libcrypto argument passes the libcrypto version that s2n was built
with. This is needed to determine the ciphers to check in our tests. Older
libcryptos do not support some modern ciphers(ChaCha20-Poly1305). Here
are the possible values for the libcrypto option:
- openssl-1.1.0
- openssl-1.0.2
- libressl

NOTE: Our .travis.yml is already setting the S2N_LIBCRYPTO environment
      variable with one of these values. We should make sure the
      .travis.yml stays in sync with acceptable inputs to the
      integration test script(s).