Replace wire_value with s2n_cipher_suite for prefs
This change changes s2n_cipher_preferences to use an array of
s2n_cipher_suite instead of a plain buffer of IANA values.
The intention is to accommodate future changes that require information
from the s2n_cipher_suite struct prior to writing the IANA value.

As a bonus, interacting with arrays of structs is less error prone
than raw buffers