Properly display skipped unit tests
Add _WITHOUT_COUNT macros for EXPECT macro checks that are used
during BEGIN_TEST and END_TEST. The _WITHOUT_COUNT macros do not
increment test_count, as test_count is initialized to 0 and using
the existing EXPECT macros increment the test_count, thus breaking
the logic for determining if tests were skipped.

It is expected that as further checks are added to BEGIN_TEST and
END_TEST, _WITHOUT_COUNT macros are added as needed to preserve
the SKIPPED test behavior.