Force EVP MD5_SHA1 hash size, remove fips_cleanup
With the EVP_DigestFinal_ex API, the size of the digest is added as
an argument. For the MD5_SHA1 case, the supplied total digest size
needs to be split up for both the MD5 and the SHA1 calls to the EVP API.

Remove s2n_fips_cleanup as s2n_fips_init will always set s2n_fips_mode
to 0 when FIPS_mode is not enabled.

Set hash_impl on s2n_hash_reset (required when in FIPS mode).