Fail Travis fuzz job on any single fuzz test failure
Currently, Travis' job status for fuzz tests is based on the return
code of the final fuzz test. With this change, the fuzz tests will
count the number of failures, ensuring that all fuzz tests are run,
and will exit with a return code of the failure count. Any single
fuzz test failure will cause the job to fail.