Add a Travis build dimension and integ test support for s2n in FIPS mode
- Add a Travis build dimension to run s2n unit tests and s2n integration
tests with s2nd in FIPS mode with OpenSSL 1.0.2. It uses the
S2N_TEST_IN_FIPS_MODE environment variable to enter FIPS mode for tests.

- A new install script fetches the latest version of OpenSSL 1.0.2 as well
as version 2.0.13 of the OpenSSL FIPS cryptographic module to build a
FIPS-capable libcrypto.

- A flag is added to s2nd to enter FIPS mode prior to s2n_init. It is
optional and does not require an argument. If it is used and FIPS mode
cannot be entered, s2nd fails to start.

- Support is added to the s2n_handshake integration tests to support FIPS
TLS protocol version and cipher restrictions.

- s2n has a new "test_all_fips" cipher preference list which contains  all
FIPS-compliant ciphers to use when testing with FIPS.