Remove FIPS-specific TLS digest preferences
As per NIST Special Publication 800-52 Revision 1, TLS 1.0 and TLS 1.1
may continue to use SHA-1 for digest preferences:

In TLS versions 1.0 and 1.1, DHE and ECDHE cipher suites use SHA-1 for
signature generation on the ephemeral parameters (including keys)
in the ServerKeyExchange message. SP800-131A states that the use of SHA-1
for digital signature generation is disallowed after 2013. Due to the
random nature of the ephemeral keys, third party cannot cause effective
collision.  The server and client do not have anything to gain by causing
collision for the connection.  Due to client random and server random,
the server, the client, or a third party cannot use a colliding set of
messages to masquerade as client or server in future connections.  Any
modification to the parameters by a third party during the handshake will
ultimately result in a failed connection. Because of these reasons, SHA-1
is allowed for generating digital signatures on ephemeral parameters in TLS.