Allow MD5 digests to be used in FIPS mode for PRF hash states
MD5 digests can only be used in FIPS mode for the PRF as per the
TLS 1.0 and 1.1 RFCs. This is confirmed by the following text from
NIST Special Publication 800-52 Revision 1:

Note that the TLS 1.1 pseudorandom function (PRF) uses MD5 and
SHA-1 in parallel so that if one hash function is broken, security
is not compromised.  While MD5 is not an Approved algorithm, the
TLS 1.1 PRF is specified as acceptable in FIPS140Impl and SP800-135.
Note that in TLS 1.1, use of SHA-1 is found acceptable for specific
cases of signing ephemeral keys and for signing for client
authentication.  This is acceptable due the fact that a third party
cannot cause a collision that is not detected and client and server
cannot exploit the collision they can cause as further explained in
a footnote in Section 3.3.1.

There were two hash states that
did not have the FIPS MD5 override flag set, causing TLS 1.0 and
1.1 handshakes to fail.