Use connection server name for default x509 hostname verification (#708)
Unless s2n_verify_host_fn is overridden, s2n will utilize connection
server_name to verify hostname on certificate following the rules from
RFC6125 6.4., with exception of 6.4.2. "Internationalized Domain Names"
and 6.4.3. p 3. "wildcard character is not the only character of the
label" for simplicity.

If server_name is not set on a connection, s2n will fail validation.

Closes #698