Fix s2n client accepting unpreferred sigalgs in ske
Previously when calling s2n_get_signature_hash_pair_if_supported in
server key exchange we accepted sigalg if result was non-zero and
rejected sigalg if result was 0. However the only 2 values that function
returns are -1 if match wasn't found and 1 if it was found. As result we
always accepted any sigalg in ske.

This change adds GUARDs around s2n_get_signature_hash_pair_if_supported
call and makes function return 0 in case of success.