Move selected cert_chain to s2n_handshake_parameters
Previously this field lived in the s2n_crypto_parameters which is a very
bloated struct that should focus on holding record security state.

This change is a part of an iterative refactor where we'd like
to move negotiated handshake parameters(cipher, certificate, sigalgs)
to the s2n_handshake_parameter struct.