Fixed a rare crash issue in tls
There are some cases that s2n_free() would free the memory but returns an error.
The utilization of GAURD() macro in caller would return the function and in
many cases that would happen before finishing all the memory release operations.
In some usages this behavior might end up in a double-free and crash. For
example, if a program calls s2n_connection_wipe() and s2n_connection_free(),
in both calls, the function s2n_connection_free_io_contexts() is getting
called. If the first call fails becaus of a munlock() issue, the second call
would try to free a memory that the first one has already free'd and that
causes a crash.
This change fixes the issue by nullifying variables that are already free'd.