Do not pass hand-crafted s2n_blob to s2n_free()
There are a few locations in code where an s2n_blob object is created and
passed to s2n_free(). For those objects, 'mlocked' fields are always set to
zero and therefore if the memory was mlocke()'d at the time of alloc() it won't
get munlock()'s at the time of free.
This change will use s2n_free_object() so that the global use_mlock variable is
taken into account when freeing the objects.
Also, a new version of s2n_array_free() is created that nullifies the caller's
reference to the array so that we won't end up in situations where GUARD()
leaves a dangling pointer.