Fixed a rare crash issue in tls (#952)
* Fixed a rare crash issue in tls

There are some cases that s2n_free() would free the memory but returns an error.
The utilization of GAURD() macro in caller would return the function and in
many cases that would happen before finishing all the memory release operations.
In some usages this behavior might end up in a double-free and crash. For
example, if a program calls s2n_connection_wipe() and s2n_connection_free(),
in both calls, the function s2n_connection_free_io_contexts() is getting
called. If the first call fails becaus of a munlock() issue, the second call
would try to free a memory that the first one has already free'd and that
causes a crash.
This change fixes the issue by nullifying variables that are already free'd.

* Do not pass hand-crafted s2n_blob to s2n_free()

There are a few locations in code where an s2n_blob object is created and
passed to s2n_free(). For those objects, 'mlocked' fields are always set to
zero and therefore if the memory was mlocke()'d at the time of alloc() it won't
get munlock()'s at the time of free.
This change will use s2n_free_object() so that the global use_mlock variable is
taken into account when freeing the objects.
Also, a new version of s2n_array_free() is created that nullifies the caller's
reference to the array so that we won't end up in situations where GUARD()
leaves a dangling pointer.