Fix receiving of signature algorithms
s2n erroneously assumed this was the last part of the message we are parsing, however there may be unrelated data to the signature algorithms in the stuffer which is not an even length.  This accounts for removing of the s2n_stuffer_data_available(in) check.  Additionally, if we are going to skip the signature algorithms extension, we should skip the bytes so as not to leave the stuffer in a bad state