Enable DRBG proof; update other SAW proofs (#984)
* commit to checkpoint record order issue

* Use different overrides for hmac_init, hmac_update, and hmac_digest.

* Remove deprecated Cryptol::Extras.

* Revert "Use different overrides for hmac_init, hmac_update, and hmac_digest."

This reverts commit 95d0ca083aa7ea052b1dd0e56f85f32329779994.

* bump saw version

* Use `(void)` for `s2n_cpu_supports_rdrand`

This clarifies to Clang that it is not a varargs function, making SAW
verification easier.

* Include extra bitcode files for DRBG proof

* Avoid using bitfields in the s2n_drbg struct

These lead to LLVM code that reads the original (maybe uninitialized)
value of a field in order to set only the relevant bits. This makes the
result dependent on an undefined value.

* Update the DRBG proof to use Crucible

This commit updates the SAW proof of DRBG to use the more modern
verification engine in SAW based on the Crucible symbolic execution
library.

* Enable DRBG proof in Travis configuration

Also include the latest SAW.

* Don't provide an override for EVP_CIPHER_CTX_reset

* Do include override for EVP_CIPHER_CTX_cleanup

* Do include override for EVP_CIPHER_CTX_init

* Put DRBG log file in a different place

* Change Z3 references to Yices references (or both)

* Use Yices for the one proof that still used Z3

* Install Yices with apt-get on Travis

* Bump SAW version on Travis

* Disable path sat checking in HMAC proof

* Disable path sat checking for SAW handshake proof

* Install Yices from a tarball

* Remove old Yices apt-get commands for Travis

* Update DRBG proof for some recent generalizations

This allows it to work with the code that’s structured to use either
AES-128 or AES-256, but assumes for the moment that it’s always using
AES-128.

* split sidetrail proofs into two travis targets

* update travis file to use sidetrail change

* fix sidetrail typo

* Disable path sat checking in HMAC proof

* Disable path sat checking for SAW handshake proof

* Install Yices from a tarball

* Bump SAW version on Travis

* Use a new function to increment DRBG counters

This function is like s2n_increment_sequence_number but explicitly
allows overflow, as this can occur safely in DRBG.

* Adapt DRBG proof to use s2n_increment_drbg_counter

This allows us to avoid most preconditions on the DRBG functions. One
precondition remains for the moment: when calling s2n_drbg_generate,
the drbg->bytes_used field must be less than half the reseed limit.

* Remove accidentally-committed file

* Remove obsolete comment from SHA256.cry

The commented-out function definition is now a primitive in Cryptol.

* Add header comments to DRBG proof files

* Use S3 URLs to fetch SAW and Yices

* Add header comment to s2n_handshake_io.saw

* Whitespace only

* Align Cryptol reseed condition with what s2n does

Fixes a bits vs. bytes disagreement.

* Remove precondition in proof of s2n_drbg_generate

* Update DRBG proof for removed global variable

* Include override for getenv in DRBG proof

* DRBG proof: require use_prediction_resistance == 1

* One more DRBG proof fix