Add cert tiebreaking callback for matching domains
This change adds a callback that is invoked when adding multiple certs
with the same domain name to an s2n_config. The callback should return
which certificate should be served for the domain name. Note that s2n
uses builtin tiebreaking based on the key type of the cert.

The default behavior of s2n without this callback set is to prefer
the *first* certificate added for a given domain name.