Add WITH_SESSION_TICKET variants with CLIENT_AUTH to state machine
While s2n does not support Session tickets with ClientAuth in server
mode due to inability to retrieve client certificate if session gets
resumed, the support for session resumption in client mode is a valid
scenario if the server has a proper handling for such scenario.

Also added integration tests with gnutls-serv to reproduce this scenario,
as gnutls-serv always sends CertificateRequest by default.