Travis update to always reinstall saw
This way it is easier to manage which saw version travis will use.
Without this you have to manually clean the cache whenever the saw version is updated.
Also it is cheap to redownload the binaries.

Maybe we can go back to caching once we have a stable saw version being used.