Prevent s2n_connection_release_buffers from corrupting record
If s2n_connection_release_buffers is called while having a partial
record, it can release the buffer dropping partial record, as a result
on next read we will end up with corrupt record and fail the connection.

This change ensure that s2n_connection_release_buffers releases buffers
only if it's safe to do so and fail otherwise.