Tolerate unexpected large legacy protocol versions
The ProtocolVersion field in the ClientHello is fixed at 0x0303 for
versions above TLS 1.2. If a strangely configured client or scanner
sends a version above this value, s2n should still try to continue the
negotiation with the highest mutually supported value. This is in line
with the suggestion in https://tools.ietf.org/html/rfc8446#section-4.2.1