Freeze maximum record protocol version to 33 (TLS 1.2).
This change is introduced with the TLS 1.3 spec
https://tools.ietf.org/html/rfc8446#section-5.1

legacy_record_version:  MUST be set to 0x0303 for all records
generated by a TLS 1.3 implementation other than an initial
ClientHello (i.e., one not generated after a HelloRetryRequest),
where it MAY also be 0x0301 for compatibility purposes.  This
field is deprecated and MUST be ignored for all purposes.
Previous versions of TLS would use other values in this field
under some circumstances.