Update handshake_io.saw to prove TLS1.3
Update the handshake_io saw proof to prove TLS1.3. The s2n method
advance_message is also updated to make the proof easier to write
and to fix a corking issue that the proof surfaced.

The rfc saw proof is updated to continue ignoring TLS1.3. Updating
that test will be a separate change.