Use switch for s2n_errno lookups
Previously we used to build a hashmap with s2n_errno mapping on s2n_init
and clean it up on s2n_cleanup. This however introduces a several
problems:

1. If we're not initialized we can't perform a fast error lookup, so we
had to introduce O(N) lookup fallback for such cases.
2. s2n_cleanup needs to be called from each thread in multi-threaded
environment, but the code to clean up error map was not thread safe.

With this change we just put all errors into a single switch statement,
which gives the benfit of compilers checking that all values are present
in a switch as well as O(1) lookups under -O2 and above as compilers
optimize switch statements with gaps into cmp and indexing memory [1][2].

[1] https://godbolt.org/z/FKeXtA
[2] https://godbolt.org/z/V7EsbB