Fix s2n_tls13_support_test and add more validaiton into version extension parsing
Stuffer blob size is the amount of memory allocated for stuffer, not the
amount of written data. Passing it as is causes extesnion parsing to go
through uninitialized memory, which sometimes causes extension parsing
to succeed, which fails the test.

While in there, changed s2n_extensions_client_supported_versions_process
to validate the size of versions list before parsing.