Allow CMake build to work with git submodules
If a parent repo adds s2n as a submodule, then CMAKE_SOURCE_DIR contains the
directory of the parent repo, not the s2n directory.

We instead want the directory that contains the current CMAKELists.txt file,
which will be present in the CMAKE_CURRENT_LIST_DIR env variable.