Zero memory as we release it
dsn noticed that when we realloc to a blob to a smaller size, we'll
fail to zero all of the memory. I don't think there are any cases of
this in s2n, but let's fix it for certain.

This change adds a blob function to create a slice, and then uses that
slice to zero the memory as we release it.