Add callbacks for memory handling (#1763)
* Brute-force comparison test for s2n_constant_time_equals

OpenSSL recently found an interesting security issue: crypto_memcmp
was comparing only the least significant bit on HPUX due to an
incorrect assembly implementation.

Our existing positive and negative test cases would not find an issue
like this (especially in the MSB).

This commit adds a test that a full brute-force comparison between
all possible 256 * 256 byte values.

* Remove trailing whitespace

* Use 17 bytes of space-left for 128-byte block hash

Performing a constant number of compression rounds is a mitigation
against Lucky13 timing attacks. s2n was not performing a constant
number of compression rounds for SHA284 based ciphersuites, as
the block size is larger 17 bytes.

The impact of this issue is categorized as "NONE" as s2n does not
use any SHA384 ciphersuites and one is included for compatibility
testing in our test suites.

This commit fixes the issue. The issue was discovered twice
independently. Our own Daniel Schwartz-Narbonne discovered
it in the course of developing SideTrail, a tool for automated
formal anaylsis of Side Channels in C code. See:

https://github.com/awslabs/s2n/tree/master/tests/sidewinder

Eyal Ronen, Kenneth G. Paterson, and Adi Shamir also discovered
it in the course of their research into further Lucky13 issues
in several implementations of TLS/SSL. See:

https://eprint.iacr.org/2018/747.pdf

* Use a ternary operator for brevity

Simplify the hash block size selection

* Implement TLS1.3 hkdf_label_expand function

This commit implements the TLS1.3 HKDF-Label-Expand function that forms the
basis of the TLS1.3 PRF and Key schedule.

The function is documented in section 7.1 of RFC8446:

    https://tools.ietf.org/html/rfc8446

and included in this commit are tests that use the test vectors from:

    https://datatracker.ietf.org/doc/draft-ietf-tls-tls13-vectors/?include_text=1

Subsequent commits will implement the PRF in full, and the corresponding
unit tests will grow to resemble the existing ones for the previous PRFs. For
now, this change gets the basis function implemented and shows that our
implementation matches the reference test vectors.

* Implement an RFC5952 inet_ntop

To verify certifcates that use an IPv6 address as their Subject Alternate Name,
we need to canonicalize IPv6 addresses regardless of platform. This change
implements an RFC5952 compliant inet_ntop function that will work the same
regardless of the local libc.

* Add missing header for s2n_inet_ntop

* Merge

* Merge differences from upstream

* Add SCRAM

Add a directory, readme and intro image for scram.

* Paragraph break

* Add a link to Steve's talk

* Fix typos

* Update the image to a etter resolution

* Integrate Shay's more precise statement

* Add a scram example image

* Be more precise

* More precision on padding

* much safer -> generally much safer

* Rounded up if needed

* Add callbacks for memory handling

This change adds a callback mechanism to s2n_mem that allows a caller
application to add its own memory handling.

The change includes documentation and a test case that adds a memory
handler that counts a histogram of allocation sizes.

* Wrap a long line and use malloc

1. wrap a long line that was breaking our builds.

2. Use malloc rather than calloc, and change our callback to match.
   On inspection s2n has no need for memory to be zero initialized.
   Modify the test case to ensure that memory is not zero initialized
   as proof.

* mlock the allocated memory

FIX: mlock the allocated memory rather than a pointer to it

* Remove mlock from the Sidetrail stubs

We've never been mlock()ing memory in the sidetrail tests. This
change removes the unreachable code-paths and references to the
now deleted blob->mlocked member.

* calloc -> malloc

* Add mlock/no_mlock memory callbacks

This change splits out our built-in memory callbacks into 2
variants. One that mlocks/madvise's and one that doesn't. This
avoids branches at run-time.

* Add mis-use resistance to s2n_mem callbacks

This change adds a guard variable to s2n_mem.c to ensure that
callbacks can not be set if s2n has already been initialized.

s2n_mem public funtions now also check that initalization
has happened.

Seperately, this change switches to "!= 0" for detecting system
call failure, rather than "< 0".

* Zero memory as we release it

dsn noticed that when we realloc to a blob to a smaller size, we'll
fail to zero all of the memory. I don't think there are any cases of
this in s2n, but let's fix it for certain.

This change adds a blob function to create a slice, and then uses that
slice to zero the memory as we release it.

* Don't zero memory if we needn't

* Spacing change

* Revert "Don't zero memory if we needn't"

This reverts commit c6aeab69b16ddbdeb372efb9c81c4debe3345a4a.

* Revert "Zero memory as we release it"

This reverts commit 9961f0c8cd4a22618e1db9be86ab25a372efcfe5.

* Revert "Add mis-use resistance to s2n_mem callbacks"

This reverts commit 81954ce783fb0ccc4db6caed1b1431651161290a.

* Add s2n_init() to the PEM fuzz test

* Revert "Revert "Add mis-use resistance to s2n_mem callbacks""

This reverts commit 9eb222116637a47287ade7557b5dc6cb6145efb4.

* Revert "Revert "Zero memory as we release it""

This reverts commit 7735ef0ded9c10ab92a28b18aabbc46716747206.

* Revert "Revert "Don't zero memory if we needn't""

This reverts commit 0fca6fa48903de8e63ca9ddfdc22877573c5d6c4.

* Add init() and clenaup()

* Avoid a needless munlock and alloc

* Add unit tests for blob

* hello_world -> world

* Remove references to blob.mlocked

* Use EXPECT_EQUAL for memcmp