Allow cbmc-batch.yaml to be clobbered at CI time
This commit replaces all cbmc-batch.yaml files with a
syntactically-incorrect YAML document. It also adds a script,
prepare.py, which is recognized by the CBMC CI system and which will be
executed before running any of the proofs. prepare.py replaces all of
the malformed YAML files with the CI parameters, which are specified in
each proof's Makefile.

This commit is intended to prevent YAML files from becoming stale with
respect to the Makefile. The source-of-truth for CI parameters is in the
Makefile, but it is possible to update the Makefile and forget to update
the YAML files. By unconditionally regenerating the YAML files on every
CI run, we guarantee that the YAML files contain up-to-date information
from the Makefile. We are still committing the YAML files, even though
they are generated, because the CBMC CI system identifies proof
directories as those containing cbmc-batch.yaml files.