Fix memory leak in STK
Previously when we received ServerHello and session wasn't resumed, we
set the conn->client_ticket.size to 0 instead of freeing the blob, later
we reused this blob through calling s2n_realloc on it, which previously
looked on size field instead of allocated to figure if blob needs to be
released, causing memory leak.

This change fixes s2n_realloc and uses s2n_free on client_ticket instead
of setting blob size to 0 directly, as the rest of codebase doesn't seem
to follow this pattern.