Add stuffer methods to handle vectors
Many places in S2N, we need to write a "vector", which means writing
the size of some data, then writing that data. This change makes this
easier, letting us go back and write the size after we've figured
out what that size is.

We can't just duplicate the stuffer to handle this because if the
stuffer resizes, it will contain different memory than the duplicate.
We can't just use raw_write, because that handles the resize problem
by erroring on resize.