Verify that the ServerHello message contains a cipher that was originally offered in the ClientHello.
Update the cipher match test to verify s2n_cipher_suite_from_wire
instead of s2n_set_cipher_as_client. This maintains the intent of
the test (unknown ciphers are ignored). But doesn't require the
additional complexity of which ciphers are available based on the
TLS library used. This complexity is what broke the test before
and required a revert commit.