Add aws-templates-for-cbmc-proofs submodule
This commit submodules in this repository and begins the process of
integrating it into the CBMC proof directory.

  https://github.com/awslabs/aws-templates-for-cbmc-proofs

That repository contains a standardized Makefile and associated data
that are used for CBMC proofs at AWS. As well as submoduling it into the
s2n repository, this commit changes some of the files that run the
proofs to be symbolic links into the submodule, and changes the
directory structure to be conformat to the templates-for-cbmc-proofs
layout. Future commits will update the proof Makefiles so that they
correctly run with the new files.