Optimize pem parsing (#2077)
When analysing performance of s2n, we notice that pem parsing takes more
CPU time than ecdhe processing. This is largely due to using single byte
read and write operations.

It is possible to safely access the data blob using the read and write
cursors while ensuring we maintain data integrity and gain a 20% speed
improvement.