Use unified Makefile template for CBMC proofs (#2080)
* Add aws-templates-for-cbmc-proofs submodule

This commit submodules in this repository and begins the process of
integrating it into the CBMC proof directory.

  https://github.com/awslabs/aws-templates-for-cbmc-proofs

That repository contains a standardized Makefile and associated data
that are used for CBMC proofs at AWS. As well as submoduling it into the
s2n repository, this commit changes some of the files that run the
proofs to be symbolic links into the submodule, and changes the
directory structure to be conformat to the templates-for-cbmc-proofs
layout. Future commits will update the proof Makefiles so that they
correctly run with the new files.

* Add s2n inclusions to CBMC proofs

* Patch proof makefiles to use makefile template

This commit changes the variable names in the individual proof Makefiles
so that they conform to the naming scheme used in the
aws-templates-for-cbmc-proofs submodule. Here is a list of the changes:

* `DEPENDENCIES` that are s2n source files are now added to
  `PROJECT_SOURCES`
* `DEPENDENCIES` that are CBMC helper functions are now added to
  `PROOF_SOURCES` and listed as under the `PROOF_SOURCE` directory
* `DEPENDENCIES` that are overrides of s2n functions are now added to
  `PROOF_SOURCES` and listed as under the `PROOF_STUB` directory
* `ENTRY` -> `HARNESS_ENTRY`
* The name of the proof harness is the value of `HARNESS_FILE`
* `HARNESS_FILE` is added to `PROOF_SOURCES`
* `UNWINDSET` -> `CBMC_UNWINDSET`
* `REMOVE_FUNCTION_BODY` -> `CBMC_REMOVE_FUNCTION_BODY`