Update saw (#2374)
* Minor updates to Cryptol specs.

The changes to tls13 make the spec forward-compatible with
pending Cryptol changes.  The DRBG_properties file also
needed some minor changes; it apparently was already several Cryptol
versions behind.

* Update SAW proofs for BIKE and SIKE to reflect interface change

* Fix update_hash_state and setup_hash_state to use readonly fields.

The C sources indicate that the first field of the `s2n_hash` struct
is a const pointer; this change makes the saw specs match.

The old definition of `update_hash_state` was actually incorrect, in
that it specified the `impl` field of the `s2n_hash` struct to be freshly
allocated every time the struct was updated. But in actuality, the
pointer value was usually copied from some other input.

A new pull request (GaloisInc/saw-script#826) makes saw-script more
strict about checking that fresh pointers are actually fresh, which
revealed the bug in the HMAC proofs. This change makes the proofs pass
the more stringent check.

* Bump SAW version.

Co-authored-by: Rob Dockins <rdockins@galois.com>
Co-authored-by: Samuel Breese <samuel@chame.co>
Co-authored-by: Brian Huffman <huffman@galois.com>
Co-authored-by: Andrei Stefanescu <andrei@stefanescu.io>