Add early data negotiation tests + misc minor fixes (#2658)
While testing negotiating the S2N handshake where early data support was requested by the client, I found a number of minor issues:

* Remove S2N_PSK_HMAC_SHA224:
No TLS1.3 cipher suite has SHA224 as a prf algorithm. Offering
it as an option to customer might be confusing. We can always
add it back if/when a TLS1.3 cipher suite supports it.

* Remove broken MIDDLEBOX_COMPAT behavior:
For now, just remove it. We'll fix it properly with
https://github.com/aws/s2n-tls/issues/2656

* Reject early data as early as possible:
We weren't rejecting early data until the EncryptedExtensions.
Although we can't accept early data before the server's response
in the EncryptedExtensions, we can reject it after the ServerHello 
f the negotiated connection parameters don't match the early
data parameters. Not doing so can lead to unexpected issues.

* Fix HRR interactions
We were checking for a hello retry handshake instead of for
a second ClientHello. We also weren't resetting the encryption
keys after a retry, leading to second ClientHellos encrypted
with the early data key.

* Fix old openssl memory leak
Old versions of Openssl can't call init twice.