Prepare s2n_config_set_psk_selection_callback to someday be async
To make s2n_config_set_psk_selection_callback someday support async,
we should set the selected PSK via methods on the opaque structure
argument, not a separate argument. This will allow the chosen PSK
to be set from outside of the original callback if necessary.

The callback can still be made async if we don't do this, but less
elegantly. We would need to retry the callback to get the result.