Only using extra warning and werror flags on Debug build type
Currently Werror and extra warnings are force into every type of build.
This is a pain point to package mantainers and cannot be deactivated
easily. Then, when a compiler or package manager that was not explicitly
tested against, for example when new compiler version is launched, the
build fails to users. Release build should not fail to users that
succesfully built after upgrading a dependency or when trying to build
to not yet supported platforms.