Fix HRR + 0RTT bug
After a retry, we can receive early data before we receive
the new ClientHello. During that time, there is no encryption
expected, so encrypted early data is treated as plaintext
records. Usually that leads to unknown record types, which we
ignore, but rarely it produces a malformed record of
a known type.

To fix this, we should never allow decryption (or encryption)
of ApplicationData with the null cipher. Even outside of early
data, it's a nice sanity check.