Make digest_allow_md5_for_fips proof UID unique
This commit fixes the UID for the
s2n_digest_allow_md5_for_fips_boringssl_awslc CBMC proof to be unique.
Previously, the UID was mistakenly set to s2n_digest_allow_md5_for_fips,
which is shared with another proof of that name. This has had the
consequence that all of the proof jobs for both those proofs have been
lumped under the same pipeline.

A check for this sort of error will be posted as a separate PR.